# packages/api-client/docs/http.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoutureCast HTTP API",
    "version": "0.1.0",
    "description": "Canonical HTTP contracts shared across API, web, mobile, and tests."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UnauthorizedHttpError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "enum": [
              401
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "error": {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "additionalProperties": false
      },
...
      "CreateOutfitCapsuleInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "occasions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "work",
                "casual",
                "formal",
                "sport",
                "travel",
                "evening",
                "outdoor",
                "home"
              ]
            },
            "minItems": 1,
            "maxItems": 8
          },
          "garmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "minItems": 2,
            "maxItems": 10
          },
          "isFavorite": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "occasions",
          "garmentIds"
        ],
        "additionalProperties": false
      },
      "OutfitCapsuleResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 1
              },
              "ownerUserId": {
                "type": "string",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 60
              },
              "description": {
                "type": [
                  "string",
                  "null"
                ],
                "maxLength": 280
              },
              "occasions": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "work",
                    "casual",
                    "formal",
                    "sport",
                    "travel",
                    "evening",
                    "outdoor",
                    "home"
                  ]
                },
                "minItems": 1,
                "maxItems": 8
              },
              "isFavorite": {
                "type": "boolean"
              },
              "revision": {
                "type": "integer",
                "minimum": 0
              },
              "availabilityStatus": {
                "type": "string",
                "enum": [
                  "ready",
                  "needs_repair"
                ]
              },
              "unavailableGarmentCount": {
                "type": "integer",
                "minimum": 0
              },
              "garments": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "category": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "top",
                        "bottom",
                        "outerwear",
                        "dress",
                        "shoes",
                        "accessory",
                        null
                      ]
                    },
                    "material": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "cotton",
                        "wool",
                        "linen",
                        "leather",
                        "denim",
                        "fleece",
                        "synthetic",
                        "down",
                        "silk",
                        null
                      ]
                    },
                    "comfortRange": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "cold",
                        "cool",
                        "mild",
                        "warm",
                        "hot",
                        null
                      ]
                    },
                    "imageAccess": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "url": {
                          "type": "string",
                          "format": "uri"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "url",
                        "expiresAt"
                      ],
                      "additionalProperties": false
                    },
                    "availabilityStatus": {
                      "type": "string",
                      "enum": [
                        "ready",
                        "needs_repair"
                      ]
                    },
                    "garmentOrder": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9
                    }
                  },
                  "required": [
                    "id",
                    "category",
                    "material",
                    "comfortRange",
                    "imageAccess",
                    "availabilityStatus",
                    "garmentOrder"
                  ],
                  "additionalProperties": false
                }
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "ownerUserId",
              "name",
              "description",
              "occasions",
              "isFavorite",
              "revision",
              "availabilityStatus",
              "unavailableGarmentCount",
              "garments",
              "createdAt",
              "updatedAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
      "FavoriteOutfitCapsuleInput": {
        "type": "object",
        "properties": {
          "isFavorite": {
            "type": "boolean"
          }
        },
        "required": [
          "isFavorite"
        ],
        "additionalProperties": false
      },
      "CapsuleBadRequestError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "enum": [
              400
            ]
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "enum": [
              "Bad Request"
            ]
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "additionalProperties": false
      },
      "CapsuleForbiddenError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "message": {
            "type": "string",
            "enum": [
              "GUARDIAN_READ_ONLY",
              "GUARDIAN_CONSENT_REQUIRED"
            ]
          },
          "error": {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/wardrobe/{ownerUserId}/capsules": {
      "post": {
        "tags": [
          "wardrobe"
        ],
        "summary": "Create a new outfit capsule for an owner wardrobe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique owner user ID."
            },
            "required": true,
            "description": "Unique owner user ID.",
            "name": "ownerUserId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutfitCapsuleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent creation replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitCapsuleResponse"
                }
              }
            }
          },
          "201": {
            "description": "Capsule created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitCapsuleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid capsule creation payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleBadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Guardian consent required or guardian read-only access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Owner user not found or unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleNotFoundError"
                }
              }
            }
          },
          "409": {
            "description": "Ineligible garments or idempotency key reused with different payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleConflictError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "wardrobe"
        ],
        "summary": "List and filter outfit capsules for an owner wardrobe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique owner user ID."
            },
            "required": true,
            "description": "Unique owner user ID.",
            "name": "ownerUserId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "work",
                "casual",
                "formal",
                "sport",
                "travel",
                "evening",
                "outdoor",
                "home"
              ]
            },
            "required": false,
            "name": "occasion",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "required": false,
            "name": "isFavorite",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "required": false,
            "name": "garmentId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "cold",
                "cool",
                "mild",
                "warm",
                "hot",
                null
              ]
            },
            "required": false,
            "name": "comfortRange",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "maximum": 10000,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of outfit capsules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitCapsuleListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleBadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Owner user not found or unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleNotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wardrobe/{ownerUserId}/capsules/{capsuleId}": {
      "get": {
        "tags": [
          "wardrobe"
        ],
        "summary": "Get details of a specific outfit capsule",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique owner user ID."
            },
            "required": true,
            "description": "Unique owner user ID.",
            "name": "ownerUserId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique outfit capsule ID."
            },
            "required": true,
            "description": "Unique outfit capsule ID.",
            "name": "capsuleId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Outfit capsule details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitCapsuleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Capsule or owner not found or unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleNotFoundError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "wardrobe"
        ],
        "summary": "Update metadata or constituent garments of an outfit capsule",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique owner user ID."
            },
            "required": true,
            "description": "Unique owner user ID.",
            "name": "ownerUserId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique outfit capsule ID."
            },
            "required": true,
            "description": "Unique outfit capsule ID.",
            "name": "capsuleId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "if-match",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutfitCapsuleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Capsule updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitCapsuleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleBadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Guardian read-only or consent forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Capsule or owner not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleNotFoundError"
                }
              }
            }
          },
          "409": {
            "description": "Ineligible garments.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleConflictError"
                }
              }
            }
          },
          "412": {
            "description": "Stale ETag / capsule revision mismatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsulePreconditionFailedError"
                }
              }
            }
          },
          "428": {
            "description": "If-Match header required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsulePreconditionRequiredError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "wardrobe"
        ],
        "summary": "Delete an outfit capsule and its join records",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique owner user ID."
            },
            "required": true,
            "description": "Unique owner user ID.",
            "name": "ownerUserId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique outfit capsule ID."
            },
            "required": true,
            "description": "Unique outfit capsule ID.",
            "name": "capsuleId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "if-match",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "Capsule deleted successfully."
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Guardian read-only or consent forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Capsule or owner not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleNotFoundError"
                }
              }
            }
          },
          "412": {
            "description": "Stale ETag / capsule revision mismatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsulePreconditionFailedError"
                }
              }
            }
          },
          "428": {
            "description": "If-Match header required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsulePreconditionRequiredError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wardrobe/{ownerUserId}/capsules/{capsuleId}/favorite": {
      "patch": {
        "tags": [
          "wardrobe"
        ],
        "summary": "Set favorite status of an outfit capsule",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique owner user ID."
            },
            "required": true,
            "description": "Unique owner user ID.",
            "name": "ownerUserId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "Unique outfit capsule ID."
            },
            "required": true,
            "description": "Unique outfit capsule ID.",
            "name": "capsuleId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "if-match",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteOutfitCapsuleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Capsule favorite status updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitCapsuleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleBadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Guardian read-only or consent forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "Capsule or owner not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsuleNotFoundError"
                }
              }
            }
          },
          "412": {
            "description": "Stale ETag / capsule revision mismatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsulePreconditionFailedError"
                }
              }
            }
          },
          "428": {
            "description": "If-Match header required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsulePreconditionRequiredError"
                }
              }
            }
          }
        }
      }
    }
  },
...
}